requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.9
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
